+++ src/components/cardInfo/CardExpiryDate.tsx
@@ -2,10 +2,10 @@ import {
   DATE_ERROR_MESSAGE,
   MONTH_ERROR_MESSAGE,
   YEAR_ERROR_MESSAGE,
-} from '../constants/messages.ts';
-import { isMonthError, isYearError } from '../utils/error.ts';
-import { isNumericInput } from '../utils/validate.ts';
-import type { CardExpiry } from '../types/cardStausTypes';
+} from '../../constants/messages.ts';
+import { isMonthError, isYearError } from '../../utils/error.ts';
+import { isNumericInput } from '../../utils/validate.ts';
+import type { CardExpiry } from '../../types/cardStausTypes.ts';
 import { useRef } from 'react';
 
 type CardExpiryDateProps = {
