+++ src/components/cardInfo/CardNumber.tsx
@@ -1,8 +1,8 @@
-import { CARD_ERROR_MESSAGE } from '../constants/messages.ts';
-import type { CardStatus } from '../types/cardStausTypes.ts';
+import { CARD_ERROR_MESSAGE } from '../../constants/messages.ts';
+import type { CardStatus } from '../../types/cardStausTypes.ts';
 import { useRef } from 'react';
-import { getCardNumberGroups } from '../utils/cardBrand.ts';
-import { isNumericInput } from '../utils/validate.ts';
+import { getCardNumberGroups } from '../../utils/cardBrand.ts';
+import { isNumericInput } from '../../utils/validate.ts';
 
 type CardNumbersProps = {
   cardStatus: CardStatus;
