+++ src/stories/CardInput.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from '@storybook/react-vite';
 import { MemoryRouter } from 'react-router-dom';
 import { expect, fn, userEvent, within } from 'storybook/test';
 
-import CardInput from '../components/CardInput';
+import CardInput from '../components/userCard/CardInput';
 import { useRegisterCardForm } from '../hooks/useRegisterCardForm';
 
 const meta = {
